Update Code
I am trialing different methods of printing the grids and how i store the values in the array. In this version the points on the grid are given a 'state', in this case hidden. The program then checks the state and prints it out. as there are no bombs and there is no player interaction the board is quite simple.